raspberrypi/requirements.txt
playsound
AWSIoTPythonSDK
paho-mqtt